Automatically install dependencies when building
When running `inko build` or related commands, the necessary
dependencies as specified in inko.pkg are downloaded automatically. This
removes the need for running `inko pkg sync` manually, though one can
still do so if needed (e.g. so they can build their project without an
internet connection).

This fixes https://github.com/inko-lang/inko/issues/722.

Changelog: added
