+++ compiler/src/pkg.rs
@@ -1,2 +1,5 @@
+pub mod git;
 pub mod manifest;
+pub mod sync;
+pub mod util;
 pub mod version;
