+++ inko/Cargo.toml
@@ -13,3 +13,4 @@ types = { path = "../types" }
 ureq = { version = "^2.9", default-features = false, features = ["tls"] }
 flate2 = "^1.0"
 tar = { version = "^0.4", default-features = false }
+blake3 = "^1.5"
