--- inko/src/pkg/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod git;
-pub(crate) mod util;
